backend/requirements.txt
Flask>=3.0,<4
Flask-Cors>=4.0,<5
Flask-Migrate>=4.0,<5
Flask-SQLAlchemy>=3.1,<4
Flask-WTF>=1.2,<2
SQLAlchemy>=2.0,<3
alembic>=1.13,<2
azure-identity>=1.15,<2
gunicorn>=22.0,<24
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
requests>=2.31,<3
